Rename feature package
