+++ app/src/main/java/app/placesautocomplete/remoterepository/RemoteAutocompleteRepository.kt
@@ -1,0 +1,8 @@
+package app.placesautocomplete.remoterepository
+
+import app.placesautocomplete.remoterepository.model.AutocompleteResponse
+
+interface RemoteAutocompleteRepository {
+
+    suspend fun getPlaces(userInput: String): AutocompleteResponse
+}
