--- app/src/main/java/app/propertyad/remoterepository/RemoteAutocompleteRepository.kt
@@ -1,8 +1,0 @@
-package app.propertyad.remoterepository
-
-import app.propertyad.remoterepository.model.AutocompleteResponse
-
-interface RemoteAutocompleteRepository {
-
-    suspend fun getPlaces(userInput: String): AutocompleteResponse
-}
